Performance experiment/requirements.txt
absl-py==2.1.0
accelerate==0.23.0
affine==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
autograd==1.7.0
autograd-gamma==0.5.0
babel==2.16.0
beautifulsoup4==4.12.3
black==21.4b2
bleach==6.2.0
blosc2==2.5.1
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1687884021435/work
cachetools==5.5.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1734380492396/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725571112467/work
cfgv==3.4.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1733218092148/work
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
csbdeep==0.7.4
cycler==0.12.1
Cython==3.0.2
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
deepspeed==0.10.3
defusedxml==0.7.1
detectron2 @ git+https://github.com/MaureenZOU/detectron2-xyz.git@42121d75e10d9f858f3a91b6a39f5722c02868f0
diffdist==0.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1733240801289/work
flake8==6.0.0
flake8-html==0.4.3
flatbuffers==24.12.23
fonttools==4.55.3
formulaic==1.1.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.4.0
genbadge==1.1.0
geojson==3.0.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1733462549337/work
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1733298745555/work
h5py==3.9.0
histolab==0.6.0
hjson==3.1.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1733299205993/work
huggingface-hub==0.27.1
hydra-core==1.3.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1733298771451/work
identify==2.6.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imagecodecs @ file:///home/conda/feedstock_root/build_artifacts/imagecodecs_1674483056579/work
imageio==2.35.1
imbalanced-learn==0.12.4
imblearn==0.0
importlib-metadata==6.11.0
importlib_resources==6.4.5
infinibatch==0.1.1
interface-meta==1.3.0
iopath==0.1.9
ipykernel==6.29.5
ipython==8.18.1
ipython-genutils==0.2.0
isodate==0.7.2
isoduration==20.11.0
jax==0.4.18
jedi==0.19.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1734823942230/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1733736026804/work
json-tricks==3.17.3
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.0.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.12.0
kiwisolver==1.4.7
kornia==0.7.0
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy-loader_1733636780672/work
libclang==18.1.1
lifelines==0.30.0
llvmlite==0.43.0
Mako==1.2.4
Markdown==3.7
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
matplotlib==3.7.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.0
ml_dtypes==0.5.0
mpi4py==4.0.1
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mup==1.0.0
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
ndindex==1.9.2
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1698504735452/work
nibabel==5.3.2
ninja==1.11.1.1
nltk==3.8.1
nodeenv==1.9.1
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.2
numpy==1.23.0
oauthlib==3.2.2
omegaconf==2.3.0
open_clip_torch==2.26.1
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
openslide-python @ file:///home/conda/feedstock_root/build_artifacts/openslide-python_1667681065457/work
opt_einsum==3.4.0
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandarallel==1.6.5
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.3.6
portalocker==2.10.1
pre-commit==3.0.3
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycodestyle==2.10.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==1.10.18
pydicom==2.4.4
pydot==3.0.1
pyflakes==3.0.1
Pygments==2.18.0
PyJWT==2.6.0
pyparsing==3.2.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-json-logger==3.2.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1733215667876/work
PyWavelets==1.6.0
PyYAML==6.0
pyzmq==26.2.0
qudida==0.0.4
rasterio==1.3.5.post1
rdflib==7.1.2
referencing==0.35.1
regex==2023.10.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.0
schema==0.7.5
scikit-image==0.19.3
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1696574864087/work
scipy==1.8.1
seaborn==0.12.2
Send2Trash==1.8.3
sentencepiece==0.1.99
sentry-sdk==2.19.2
setproctitle==1.3.4
Shapely==1.8.5.post1
SimpleITK==2.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.6
SQLAlchemy==2.0.17
stack-data==0.6.3
stardist==0.8.5
sympy==1.13.1
tables==3.9.2
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.1
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1714400101435/work
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1692120125985/work
timm==1.0.14
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torch-geometric==2.6.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
triton==3.1.0
types-python-dateutil==2.9.0.20241206
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1733235305708/work
ujson==5.8.0
uri-template==1.3.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1734859416348/work
virtualenv==20.28.1
vision-datasets==0.2.2
wandb==0.15.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.14.1
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.23.0
